+++ template/test/before.ts
@@ -3,7 +3,7 @@ import { mongodb } from '../src/components';
 describe('connectDB', () => {
     //#module mongodb
     before((done) => {
-        mongodb.connection.once('open', () => {
+        mongodb.once('open', () => {
             console.log('mongodb connected');
             done();
         });